+++ interpreter/builtin_sequence.h
@@ -1,13 +1,13 @@
 /*
- * tré – Copyright (c) 2005–2008 Sven Michael Klose <pixel@copei.de>
+ * tré – Copyright (c) 2005–2008,2013 Sven Michael Klose <pixel@copei.de>
  */
 
 #ifndef TRE_SEQUENCE_H
 #define TRE_SEQUENCE_H
 
 struct tre_sequence_type {
-     void    (*set) (treptr, ulong, treptr);
-     treptr  (*get) (treptr, ulong);
+     void    (*set) (treptr, size_t, treptr);
+     treptr  (*get) (treptr, size_t);
      size_t  (*length) (treptr);
 };
