+++ interpreter/error.h
@@ -1,5 +1,5 @@
 /*
- * tré – Copyright (c) 2005–2008 Sven Michael Klose <pixel@copei.de>
+ * tré – Copyright (c) 2005–2008,2013 Sven Michael Klose <pixel@copei.de>
  */
 
 #ifndef TRE_ERROR_H
@@ -10,6 +10,6 @@ extern treptr treerror (treptr, const char *msg, ...);
 extern void   treerror_norecover (treptr, const char *msg, ...);
 extern void   trewarn (treptr, const char *msg, ...);
 
-extern const char * treerror_typename (ulong);
+extern const char * treerror_typename (size_t);
 
 #endif	/* #ifndef TRE_ERROR_H */
