+++ interpreter/io.h
@@ -1,5 +1,5 @@
 /*
- * tré – Copyright (c) 2005–2008 Sven Michael Klose <pixel@copei.de>
+ * tré – Copyright (c) 2005–2008,2013 Sven Michael Klose <pixel@copei.de>
  */
 
 #ifndef TRE_IO_H
@@ -9,13 +9,13 @@
 
 struct tre_stream {
     struct treio_ops  *ops;
-    int		putback_char;
-    int		last_char;
-    void	* detail_in;
-    void	* detail_out;
-	char  	* file_name;
-	ulong	line;
-	ulong	column;
+    int    putback_char;
+    int    last_char;
+    void * detail_in;
+    void * detail_out;
+	char * file_name;
+	size_t line;
+	size_t column;
 };
 
 struct treio_ops {
@@ -36,14 +36,14 @@ struct treio_ops {
 extern struct tre_stream  *treio_reader;
 extern struct tre_stream  *treio_console;
 
-extern ulong treio_readerstreamptr;
+extern size_t treio_readerstreamptr;
 
 extern struct tre_stream * treio_make_stream (struct treio_ops *, const char * name);
 extern void treio_free_stream (struct tre_stream *);
 extern void treio_close_stream (struct tre_stream *);
 
 extern int  treio_getc (struct tre_stream *);
-extern int  treio_getline (struct tre_stream *,char *s, ulong maxlen);
+extern int  treio_getline (struct tre_stream *,char *s, size_t maxlen);
 extern void treio_putback (struct tre_stream *);
 extern void treio_putc (struct tre_stream *, char);
 extern void treio_flush (struct tre_stream *);
